% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regressions.R
\name{boot_reg}
\alias{boot_reg}
\title{Perform a single equation regression model on the data}
\usage{
boot_reg(data, formula, label = as.character(formula), variable,
  bootR = 2000)
}
\arguments{
\item{data}{dataframe for the regression model}

\item{formula}{formula for the regression model}

\item{label}{character string with the name of the test}

\item{variable}{character string with the name of the variable that tests for
the interdependency}

\item{bootR}{integer value for the number of bootstrap repetitions}
}
\description{
Perform a single equation regression model on the data
}
