% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{check_numeric}
\alias{check_numeric}
\title{check whether an object is numeric and of a predefined length}
\usage{
check_numeric(input, len, interval = NULL)
}
\arguments{
\item{input}{The object to check}

\item{len}{The desired length of the object}

\item{interval}{The interval the inputs should be in}
}
\description{
check whether an object is numeric and of a predefined length
}
