% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Perform a simulation}
\usage{
run_simulation(data_params, test_params, sim_params = list(nsim = 100,
  mc_cores = 1, boot = FALSE, bootR = 2000))
}
\arguments{
\item{data_params}{List of parameters to be used to simulate data.}

\item{test_params}{List of with formula, variable, and label to use as a
test for the complementarity}

\item{sim_params}{List of simulation specific parameters.
\describe{
\item{\code{nsim}}{Integer for the number of replications in the
simulation}
\item{\code{mc_cores}}{Integer for the number of cores to be used in the
 parallel}
\item{\code{boot}}{Boolean whether the bootstrap version is used
to estimate standard errors.}
\item{\code{bootR}}{The number of bootstrap replications}
}}
}
\description{
Perform a simulation
}
