% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regressions.R
\name{single_reg}
\alias{single_reg}
\title{Perform a single equation regression model on the data}
\usage{
single_reg(data, formula, label = as.character(formula), variable)
}
\arguments{
\item{data}{dataframe for the regression model}

\item{formula}{formula for the regression model}

\item{label}{character string with the name of the test}

\item{variable}{character string with the name of the variable that tests for
the interdependency}
}
\description{
Perform a single equation regression model on the data
}
